make  coupledSDEs an extention
